Revert `_ANALYTICS_` vars forwarding — It's not helpful
Contrary to what I expected, the logs from a build with these values had
_less_ information, not more.
See
https://buildkite.com/automattic/testpress-apple/builds/37#0181e9ea-4e69-4bc0-805c-03f603b63c03/672-767

In hindsight, maybe that's not surprising? 🤔 The test collector library gives
precedence to the `_ANALYTICS_` values. I assumed they would have been
available in the Buildkite environment, but maybe (apparently from the
logs) they aren't, so it's possible they are meant as debug overrides,
not as values to be used by consumers.